+++ src/store/modules/loans.js
@@ -5,6 +5,7 @@ const apiName = 'loans'
 const state = {
     dataLoan: [],
     statusLoan: {},
+    dataLoanByDept : [],
 }
 
 const actions = {
@@ -12,17 +13,26 @@ const actions = {
         const res = await httpCommons.post(apiName, data)
         // console.log(res)
         commit('SET_STATUS_LOAN', res)
+    },
+    
+    async getTotalLoanPerDepartment({commit}){
+        const res = await httpCommons.get(apiName+'/totaLoanByDepartment')
+        commit('SET_DATA_LOAN_BY_DEPT', res.data)
     }
 }
 
 const mutations = {
     SET_STATUS_LOAN(state, rows){
         state.statusLoan = rows
+    },
+    SET_DATA_LOAN_BY_DEPT(state, rows){
+        state.dataLoanByDept = rows
     }
 }
 
 const getters ={
-    getStatusLoan : state => state.statusLoan
+    getStatusLoan : state => state.statusLoan,
+    getDataLoanByDept : state => state.dataLoanByDept
 }
 
 export default {